refactor: apps folder and separate user app
