--- MnA_BE/DailyInsight/sample/urls.py
@@ -1,5 +1,0 @@
-from django.urls import path
-from . import views
-
-urlpatterns = [
-]
