--- MnA_BE/DailyInsight/views.py
@@ -1,2 +1,0 @@
-# dummy file for django
-# Don't remove this file
