+++ MnA_BE/MnA_BE/settings.py
@@ -28,13 +28,28 @@ SECRET_KEY = "django-insecure-9*#xm@jz&k9%1dpl$nq0d1_&wd0n9qua=krj(ucab+(+xtil+!
 # SECURITY WARNING: don't run with debug turned on in production!
 DEBUG = True
 
+
+# CORS Settings
 ALLOWED_HOSTS = []
+CORS_ORIGIN_ALLOW_ALL = True
+CORS_ALLOW_CREDENTIALS = True
+
+CORS_ALLOW_METHODS = (
+    'DELETE',
+    'GET',
+    'POST',
+    'PUT',
+    'HEAD'
+)
+
+CORS_ALLOW_HEADERS = (
+    'content-type'
+)
 
 
 # Application definition
 
 INSTALLED_APPS = [
-    "DailyInsight.apps.DailyInsightConfig",
     "django.contrib.admin",
     "django.contrib.auth",
     "django.contrib.contenttypes",
@@ -44,6 +59,7 @@ INSTALLED_APPS = [
 ]
 
 MIDDLEWARE = [
+    "corsheaders.middleware.CorsMiddleware",
     "django.middleware.security.SecurityMiddleware",
     "django.contrib.sessions.middleware.SessionMiddleware",
     "django.middleware.common.CommonMiddleware",
