+++ MnA_BE/MnA_BE/urls.py
@@ -19,6 +19,7 @@ from django.contrib import admin
 from django.urls import path, include
 
 urlpatterns = [
-    path("DailyInsight/", include('DailyInsight.urls')),
+    path("user/", include('apps.user.urls')),
+    path("DailyInsight/", include('apps.DailyInsight.urls')),
     path("admin/", admin.site.urls),
 ]
