+++ MnA_BE/apps/DailyInsight/admin.py
