+++ MnA_BE/apps/DailyInsight/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class DailyInsightConfig(AppConfig):
+class DailyinsightConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
     name = "DailyInsight"
