+++ MnA_BE/apps/DailyInsight/models.py
@@ -1,16 +1,5 @@
 from django.db import models
-
-TINY_TEXT = 255
-LONG_TEXT = 255
-
-# Create your models here.
-class User(models.Model):
-    id = models.AutoField(primary_key=True)
-    password = models.CharField(max_length=TINY_TEXT)
-    name = models.CharField(max_length=TINY_TEXT)
-
-    # bitmask string for each tag(On / Off)
-    interests = models.CharField(max_length=LONG_TEXT)
+from MnA_BE.constants import *
 
 class PriceFluctuation(models.Model):
     recommended_date = models.DateField()
